Query DB once when App is mounted, not when routing to Gallery
